Changed the Backgrounds of the Git Stats and put languages and tools in a summary-details box
I've changed the background color of the git stats pictures as the contrast wasn't very readable. I've also put the languages, and tools in a summary-details box as the lists were getting long, and I think that it'd be better for them to see the github activity chart instead of the languages. As the former is more impressive